# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorBody {
  message: string;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  READY: 'READY',
  INCOMPLETE: 'INCOMPLETE',
  BLOCKED: 'BLOCKED',
  PROCESSING: 'PROCESSING',
  ERROR: 'ERROR',
} as const;

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  BLOCKING: 'BLOCKING',
  RISK: 'RISK',
  WARNING: 'WARNING',
} as const;

export type FollowUpStatus = typeof FollowUpStatus[keyof typeof FollowUpStatus];


export const FollowUpStatus = {
  pending: 'pending',
  sent: 'sent',
  resolved: 'resolved',
} as const;

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  skill: 'skill',
  tool: 'tool',
  status: 'status',
} as const;

export type PipelineStepState = typeof PipelineStepState[keyof typeof PipelineStepState];


export const PipelineStepState = {
  completed: 'completed',
  current: 'current',
  blocked: 'blocked',
  pending: 'pending',
} as const;

export type PipelineActor = typeof PipelineActor[keyof typeof PipelineActor];


export const PipelineActor = {
  skill: 'skill',
  tool: 'tool',
  system: 'system',
} as const;

export interface InboxSummary {
  documentsToday: number;
  readyCount: number;
  incompleteCount: number;
  blockedCount: number;
  processingCount: number;
}

export interface DocumentSummary {
  id: string;
  filename: string;
  documentType: string;
  entityName: string;
  clientName?: string;
  status: DocumentStatus;
  score: number;
  anomalyCount: number;
  createdAt: string;
  amountTtc?: number;
  regime?: string;
}

export interface DocumentListResponse {
  items: DocumentSummary[];
  total: number;
}

export interface ExtractedField {
  key: string;
  label: string;
  value: string | null;
  missing: boolean;
  group?: string;
}

export interface Anomaly {
  code: string;
  severity: Severity;
  description: string;
  impact: string;
}

export interface FollowUpMessage {
  id: string;
  documentId: string;
  documentName: string;
  recipient: string;
  severity: Severity;
  status: FollowUpStatus;
  fr: string;
  ar: string;
  createdAt: string;
  reviewed?: boolean;
}

export interface FollowUpListResponse {
  items: FollowUpMessage[];
}

export interface PipelineStep {
  id: string;
  label: string;
  state: PipelineStepState;
  actor: PipelineActor;
  actorName?: string;
  explanation: string;
  completedAt?: string | null;
}

export type SkillRunStatus = typeof SkillRunStatus[keyof typeof SkillRunStatus];


export const SkillRunStatus = {
  success: 'success',
  warning: 'warning',
  error: 'error',
  idle: 'idle',
} as const;

export interface SkillRun {
  name: string;
  role: string;
  inputSummary: string;
  outputSummary: string;
  confidence?: number | null;
  status: SkillRunStatus;
}

export type DeterministicCheckResult = typeof DeterministicCheckResult[keyof typeof DeterministicCheckResult];


export const DeterministicCheckResult = {
  passed: 'passed',
  failed: 'failed',
  warning: 'warning',
} as const;

export interface DeterministicCheck {
  name: string;
  verified: string;
  result: DeterministicCheckResult;
  rationale: string;
}

export interface ReadinessDecision {
  score: number;
  status: DocumentStatus;
  blockingFields: string[];
  warnings: string[];
  nextAction: string;
  declarationMonth?: string;
}

export type DocumentDetailAmount = {
  ht?: number;
  tva?: number;
  ttc?: number;
};

export interface DocumentDetail {
  id: string;
  filename: string;
  documentType: string;
  entityName: string;
  clientName?: string;
  status: DocumentStatus;
  score: number;
  anomalyCount: number;
  createdAt: string;
  regime?: string;
  confidence?: number | null;
  amount?: DocumentDetailAmount;
  extractedFields: ExtractedField[];
  anomalies: Anomaly[];
  timeline: PipelineStep[];
  skillsUsed: SkillRun[];
  deterministicChecks: DeterministicCheck[];
  readiness: ReadinessDecision;
  followups?: FollowUpMessage[];
}

export interface AgentActivity {
  id: string;
  type: ActivityType;
  title: string;
  detail: string;
  createdAt: string;
  severity?: Severity;
  skillOrTool?: string;
}

export interface AgentActivityListResponse {
  items: AgentActivity[];
}

export type SkillDefinitionStatus = typeof SkillDefinitionStatus[keyof typeof SkillDefinitionStatus];


export const SkillDefinitionStatus = {
  active: 'active',
  idle: 'idle',
  error: 'error',
} as const;

export interface SkillDefinition {
  name: string;
  title: string;
  description: string;
  inputSchemaSummary: string;
  outputSchemaSummary: string;
  stage: string;
  status: SkillDefinitionStatus;
  usedWhen?: string;
}

export interface SkillListResponse {
  items: SkillDefinition[];
}

export type ToolDefinitionStatus = typeof ToolDefinitionStatus[keyof typeof ToolDefinitionStatus];


export const ToolDefinitionStatus = {
  active: 'active',
  coming_soon: 'coming_soon',
} as const;

export interface ToolDefinition {
  name: string;
  title: string;
  description: string;
  status: ToolDefinitionStatus;
  verifies?: string;
}

export interface ToolListResponse {
  items: ToolDefinition[];
}

export type AgentOverviewStatus = typeof AgentOverviewStatus[keyof typeof AgentOverviewStatus];


export const AgentOverviewStatus = {
  active: 'active',
  idle: 'idle',
  processing: 'processing',
} as const;

export interface AgentOverview {
  name: string;
  status: AgentOverviewStatus;
  documentsAnalyzed: number;
  averageReadinessScore: number;
  anomaliesThisWeek: number;
  skillsActive?: number;
  lastPipelineRun?: string;
}

export type ListDocumentsParams = {
status?: DocumentStatus;
search?: string;
sort?: ListDocumentsSort;
};

export type ListDocumentsSort = typeof ListDocumentsSort[keyof typeof ListDocumentsSort];


export const ListDocumentsSort = {
  newest: 'newest',
  oldest: 'oldest',
  score_desc: 'score_desc',
  score_asc: 'score_asc',
  anomalies_desc: 'anomalies_desc',
} as const;

export type ListFollowupsParams = {
status?: FollowUpStatus;
severity?: Severity;
language?: ListFollowupsLanguage;
};

export type ListFollowupsLanguage = typeof ListFollowupsLanguage[keyof typeof ListFollowupsLanguage];


export const ListFollowupsLanguage = {
  fr: 'fr',
  ar: 'ar',
} as const;

export type ListAgentActivityParams = {
limit?: number;
};