# lib/api-zod/src/generated/types/activityType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  skill: 'skill',
  tool: 'tool',
  status: 'status',
} as const;