# lib/api-zod/src/generated/types/agentActivity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { ActivityType } from './activityType';
import type { Severity } from './severity';

export interface AgentActivity {
  id: string;
  type: ActivityType;
  title: string;
  detail: string;
  createdAt: string;
  severity?: Severity;
  skillOrTool?: string;
}