# lib/api-zod/src/generated/types/agentOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { AgentOverviewStatus } from './agentOverviewStatus';

export interface AgentOverview {
  name: string;
  status: AgentOverviewStatus;
  documentsAnalyzed: number;
  averageReadinessScore: number;
  anomaliesThisWeek: number;
  skillsActive?: number;
  lastPipelineRun?: string;
}