# lib/api-zod/src/generated/types/deterministicCheck.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { DeterministicCheckResult } from './deterministicCheckResult';

export interface DeterministicCheck {
  name: string;
  verified: string;
  result: DeterministicCheckResult;
  rationale: string;
}