# lib/api-zod/src/generated/types/documentDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { Anomaly } from './anomaly';
import type { DeterministicCheck } from './deterministicCheck';
import type { DocumentDetailAmount } from './documentDetailAmount';
import type { DocumentStatus } from './documentStatus';
import type { ExtractedField } from './extractedField';
import type { FollowUpMessage } from './followUpMessage';
import type { PipelineStep } from './pipelineStep';
import type { ReadinessDecision } from './readinessDecision';
import type { SkillRun } from './skillRun';

export interface DocumentDetail {
  id: string;
  filename: string;
  documentType: string;
  entityName: string;
  clientName?: string;
  status: DocumentStatus;
  score: number;
  anomalyCount: number;
  createdAt: string;
  regime?: string;
  confidence?: number | null;
  amount?: DocumentDetailAmount;
  extractedFields: ExtractedField[];
  anomalies: Anomaly[];
  timeline: PipelineStep[];
  skillsUsed: SkillRun[];
  deterministicChecks: DeterministicCheck[];
  readiness: ReadinessDecision;
  followups?: FollowUpMessage[];
}