# lib/api-zod/src/generated/types/documentStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  READY: 'READY',
  INCOMPLETE: 'INCOMPLETE',
  BLOCKED: 'BLOCKED',
  PROCESSING: 'PROCESSING',
  ERROR: 'ERROR',
} as const;