# lib/api-zod/src/generated/types/documentSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { DocumentStatus } from './documentStatus';

export interface DocumentSummary {
  id: string;
  filename: string;
  documentType: string;
  entityName: string;
  clientName?: string;
  status: DocumentStatus;
  score: number;
  anomalyCount: number;
  createdAt: string;
  amountTtc?: number;
  regime?: string;
}