# lib/api-zod/src/generated/types/followUpListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { FollowUpMessage } from './followUpMessage';

export interface FollowUpListResponse {
  items: FollowUpMessage[];
}