# lib/api-zod/src/generated/types/inboxSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */

export interface InboxSummary {
  documentsToday: number;
  readyCount: number;
  incompleteCount: number;
  blockedCount: number;
  processingCount: number;
}