# lib/api-zod/src/generated/types/pipelineActor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */

export type PipelineActor = typeof PipelineActor[keyof typeof PipelineActor];


export const PipelineActor = {
  skill: 'skill',
  tool: 'tool',
  system: 'system',
} as const;