# lib/api-zod/src/generated/types/readinessDecision.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { DocumentStatus } from './documentStatus';

export interface ReadinessDecision {
  score: number;
  status: DocumentStatus;
  blockingFields: string[];
  warnings: string[];
  nextAction: string;
  declarationMonth?: string;
}