# lib/api-zod/src/generated/types/severity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  BLOCKING: 'BLOCKING',
  RISK: 'RISK',
  WARNING: 'WARNING',
} as const;