# lib/api-zod/src/generated/types/skillRun.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Declaration Copilot API
 * OpenAPI spec version: 0.2.0
 */
import type { SkillRunStatus } from './skillRunStatus';

export interface SkillRun {
  name: string;
  role: string;
  inputSummary: string;
  outputSummary: string;
  confidence?: number | null;
  status: SkillRunStatus;
}